+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 - [SeqKit v2.10.0](https://github.com/shenwei356/seqkit/releases/tag/v2.10.0) - 2025-xx-xx
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v2.10.0/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v2.10.0)
+    - `seqkit split2`:
+        - fix prefix checking when paired-end files are given. [#512](https://github.com/shenwei356/seqkit/issues/512)
     - `seqkit stat`:
         - do not compute GC content and N's for protein sequence. [#497](https://github.com/shenwei356/seqkit/issues/497)
     - `seqkit grep`: